breaking to abstract change, finish criticality work
