+++ lib/graphql/schema_comparator/changes/non_breaking.rb
@@ -6,15 +6,11 @@ module GraphQL
 
         def initialize(type)
           @type = type
-          @breaking = false
+          @criticality = Changes::Criticality.non_breaking
         end
 
         def message
           "Type `#{type.name}` was added"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -23,15 +19,11 @@ module GraphQL
 
         def initialize(directive)
           @directive = directive
-          @breaking = false
+          @criticality = Changes::Criticality.non_breaking
         end
 
         def message
           "Directive `#{directive.name}` was added"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -41,15 +33,11 @@ module GraphQL
         def initialize(old_type, new_type)
           @old_type = old_type
           @new_type = new_type
-          @breaking = false
+          @criticality = Changes::Criticality.non_breaking
         end
 
         def message
           "Description `#{old_type.description}` on type `#{old_type.name}` has changed to `#{new_type.description}`"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -60,15 +48,12 @@ module GraphQL
           @enum = enum
           @old_enum_value = old_enum_value
           @new_enum_value = new_enum_value
+          @criticality = Changes::Criticality.non_breaking
         end
 
         def message
           "Description for enum value `#{enum.name}.#{new_enum_value.name}` changed from " \
             "`#{old_enum_value.description}` to `#{new_enum_value.description}`"
-        end
-
-        def breaking?
-          false
         end
       end
 
@@ -76,6 +61,7 @@ module GraphQL
         attr_reader :enum, :old_enum_value, :new_enum_value, :criticality
 
         def initialize(enum, old_enum_value, new_enum_value)
+          @criticality = Changes::Criticality.non_breaking
           @enum = enum
           @old_enum_value = old_enum_value
           @new_enum_value = new_enum_value
@@ -90,29 +76,21 @@ module GraphQL
               " `#{new_enum_value.deprecation_reason}`"
           end
         end
-
-        def breaking?
-          false
-        end
       end
 
       class InputFieldDescriptionChanged < AbstractChange
         attr_reader :input_type, :old_field, :new_field, :criticality
 
         def initialize(input_type, old_field, new_field)
+          @criticality = Changes::Criticality.non_breaking
           @input_type = input_type
           @old_field = old_field
           @new_field = new_field
-          @breaking = false
         end
 
         def message
           "Input field `#{input_type.name}.#{old_field.name}` description changed"\
             " from `#{old_field.description}` to `#{new_field.description}`"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -120,18 +98,14 @@ module GraphQL
         attr_reader :old_directive, :new_directive, :criticality
 
         def initialize(old_directive, new_directive)
+          @criticality = Changes::Criticality.non_breaking
           @old_directive = old_directive
           @new_directive = new_directive
-          @breaking = false
         end
 
         def message
           "Directive `#{new_directive.name}` description changed"\
             " from `#{old_directive.description}` to `#{new_directive.description}`"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -139,19 +113,15 @@ module GraphQL
         attr_reader :type, :old_field, :new_field, :criticality
 
         def initialize(type, old_field, new_field)
+          @criticality = Changes::Criticality.non_breaking
           @type = type
           @old_field = old_field
           @new_field = new_field
-          @breaking = false
         end
 
         def message
           "Field `#{type.name}.#{old_field.name}` description changed"\
             " from `#{old_field.description}` to `#{new_field.description}`"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -159,20 +129,16 @@ module GraphQL
         attr_reader :type, :field, :old_argument, :new_argument, :criticality
 
         def initialize(type, field, old_argument, new_argument)
+          @criticality = Changes::Criticality.non_breaking
           @type = type
           @field = field
           @old_argument = old_argument
           @new_argument = new_argument
-          @breaking = false
         end
 
         def message
           "Description for argument `#{new_argument.name}` on field `#{type.name}.#{field.name}` changed"\
             " from `#{old_argument.description}` to `#{new_argument.description}`"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -180,19 +146,15 @@ module GraphQL
         attr_reader :directive, :old_argument, :new_argument, :criticality
 
         def initialize(directive, old_argument, new_argument)
+          @criticality = Changes::Criticality.non_breaking
           @directive = directive
           @old_argument = old_argument
           @new_argument = new_argument
-          @breaking = false
         end
 
         def message
           "Description for argument `#{new_argument.name}` on directive `#{directive.name}` changed"\
             " from `#{old_argument.description}` to `#{new_argument.description}`"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -200,19 +162,15 @@ module GraphQL
         attr_reader :type, :old_field, :new_field, :criticality
 
         def initialize(type, old_field, new_field)
+          @criticality = Changes::Criticality.non_breaking
           @type = type
           @old_field = old_field
           @new_field = new_field
-          @breaking = false
         end
 
         def message
           "Deprecation reason on field `#{type.name}.#{new_field.name}` has changed "\
             "from `#{old_field.deprecation_reason}` to `#{new_field.deprecation_reason}`"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -220,19 +178,15 @@ module GraphQL
         attr_reader :input_type, :old_field, :new_field, :criticality
 
         def initialize(input_type, old_field, new_field)
+          @criticality = Changes::Criticality.non_breaking
           @input_type = input_type
           @old_field = old_field
           @new_field = new_field
-          @breaking = false
         end
 
         def message
           "Input field `#{input_type.name}.#{old_field.name}` default changed"\
             " from `#{old_field.default_value}` to `#{new_field.default_value}`"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -240,19 +194,15 @@ module GraphQL
         attr_reader :directive, :old_argument, :new_argument, :criticality
 
         def initialize(directive, old_argument, new_argument)
+          @criticality = Changes::Criticality.non_breaking
           @directive = directive
           @old_argument = old_argument
           @new_argument = new_argument
-          @breaking = false
         end
 
         def message
           "Default value for argument `#{new_argument.name}` on directive `#{directive.name}` changed"\
             " from `#{old_argument.default_value}` to `#{new_argument.default_value}`"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -260,17 +210,13 @@ module GraphQL
         attr_reader :interface, :object_type, :criticality
 
         def initialize(interface, object_type)
+          @criticality = Changes::Criticality.non_breaking
           @interface = interface
           @object_type = object_type
-          @breaking = false
         end
 
         def message
           "`#{object_type.name}` object implements `#{interface.name}` interface"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -278,35 +224,28 @@ module GraphQL
         attr_reader :object_type, :field, :criticality
 
         def initialize(object_type, field)
+          @criticality = Changes::Criticality.non_breaking
           @object_type = object_type
           @field = field
-          @breaking = false
         end
 
         def message
           "Field `#{field.name}` was added to object type `#{object_type.name}`"
         end
 
-        def breaking?
-          !!@breaking
-        end
       end
 
       class DirectiveLocationAdded < AbstractChange
         attr_reader :directive, :location, :criticality
 
         def initialize(directive, location)
+          @criticality = Changes::Criticality.non_breaking
           @directive = directive
           @location = location
-          @breaking = false
         end
 
         def message
           "Location `#{location}` was added to directive `#{directive.name}`"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
 
@@ -454,123 +393,17 @@ module GraphQL
         end
       end
 
-      class InputFieldAdded < AbstractChange
-        attr_reader :input_object_type, :field, :criticality
-
-        def initialize(input_object_type, field)
-          @input_object_type = input_object_type
-          @field = field
-          @breaking = field.type.kind.non_null? ? true : false
-        end
-
-        def message
-          "Input field `#{field.name}` was added to input object type `#{input_object_type.name}`"
-        end
-
-        def breaking?
-          !!@breaking
-        end
-      end
-
-      class FieldArgumentAdded < AbstractChange
-        attr_reader :type, :field, :argument, :criticality
-
-        def initialize(type, field, argument)
-          @type = type
-          @field = field
-          @argument = argument
-          # TODO: should at least have a warning that it may still be breaking
-          @breaking = argument.type.kind.non_null? ? true : false
-        end
-
-        def message
-          "Argument `#{argument.name}: #{argument.type}` added to field `#{type.name}.#{field.name}`"
-        end
-
-        def breaking?
-          !!@breaking
-        end
-      end
-
       class DirectiveArgumentAdded < AbstractChange
         attr_reader :directive, :argument, :criticality
 
         def initialize(directive, argument)
+          @criticality = Changes::Criticality.non_breaking
           @directive = directive
           @argument = argument
-          @breaking = false
         end
 
         def message
           "Argument `#{argument.name}` was added to directive `#{directive.name}`"
-        end
-
-        def breaking?
-          !!@breaking
-        end
-      end
-
-      class InputFieldTypeChanged < AbstractChange
-        include SafeTypeChange
-
-        attr_reader :input_type, :old_input_field, :new_input_field, :criticality
-
-        def initialize(input_type, old_input_field, new_input_field)
-          @input_type = input_type
-          @old_input_field = old_input_field
-          @new_input_field = new_input_field
-          @breaking = !safe_change_for_input_value?(old_input_field.type, new_input_field.type)
-        end
-
-        def message
-          "Input field `#{input_type}.#{old_input_field.name}` changed type from #{old_input_field.type} to #{new_input_field.type}"
-        end
-
-        def breaking?
-          !!@breaking
-        end
-      end
-
-      class FieldArgumentTypeChanged < AbstractChange
-        include SafeTypeChange
-
-        attr_reader :type, :field, :old_argument, :new_argument, :criticality
-
-        def initialize(type, field, old_argument, new_argument)
-          @type = type
-          @field = field
-          @old_argument = old_argument
-          @new_argument = new_argument
-          @breaking = !safe_change_for_input_value?(old_argument.type, new_argument.type)
-        end
-
-        def message
-          "Type for argument `#{new_argument.name}` on field `#{type.name}.#{field.name}` changed"\
-            " from `#{old_argument.type}` to `#{new_argument.type}`"
-        end
-
-        def breaking?
-          !!@breaking
-        end
-      end
-
-      class DirectiveArgumentTypeChanged < AbstractChange
-        attr_reader :directive, :old_argument, :new_argument, :criticality
-
-        def initialize(directive, old_argument, new_argument)
-          @directive = directive
-          @old_argument = old_argument
-          @new_argument = new_argument
-          @breaking = false
-        end
-
-        def message
-          "Type for argument `#{new_argument.name}` on directive `#{directive.name}` changed"\
-            " from `#{old_argument.type}` to `#{new_argument.type}`"
-        end
-
-        def breaking?
-          !!@breaking
         end
       end
     end
