+++ test/lib/graphql/schema_comparator/result_test.rb
@@ -7,7 +7,7 @@ describe GraphQL::SchemaComparator::Result do
       removed_a = GraphQL::SchemaComparator::Changes::FieldRemoved.new(GraphQL::ObjectType.define(name: "A"), GraphQL::Field.define(name: "a"))
       added_a = GraphQL::SchemaComparator::Changes::FieldAdded.new(GraphQL::ObjectType.define(name: "A"), GraphQL::Field.define(name: "a"))
       result = GraphQL::SchemaComparator::Result.new([removed_z, added_a, removed_a])
-      assert_equal [removed_a, removed_z, added_a], result.changes
+      assert_equal [removed_z, removed_a, added_a], result.changes
     end
   end
 
@@ -62,7 +62,7 @@ describe GraphQL::SchemaComparator::Result do
       ])
 
       assert_equal [field_removed], result.breaking_changes
-      assert_equal [type_description_changed, field_added], result.non_breaking_changes
+      assert_equal [field_added, type_description_changed], result.non_breaking_changes
     end
   end
 end
